// src/modules/dashboard/enums/DashboardKycStep.ts
export enum DashboardKycStep {
  BASIC_INFORMATION = "BASIC_INFORMATION",
  IDENTIFICATION = "IDENTIFICATION",
  ALTERNATE_PHONE_NUMBER = "ALTERNATE_PHONE_NUMBER",
  ACCOUNT_DETAILS = "ACCOUNT_DETAILS",
  ADDRESS_VERIFICATION = "ADDRESS_VERIFICATION",
  TAX_IDENTIFICATION_NUMBER = "TAX_IDENTIFICATION_NUMBER",
  SUCCESS = "SUCCESS",
}